% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tlasso-package.r
\docType{package}
\name{Tlasso}
\alias{Tlasso}
\title{Non-Convex Optimization and Statistical Inference for Sparse Tensor Graphical Models}
\description{
An optimal alternating optimization algorithm for estimation of precision matrices of sparse tensor graphical models, and an efficient inference procedure for support recovery of the precision matrices.
}
\details{
\tabular{ll}{
  Package: \tab Tlasso \cr
  Type: \tab Package \cr
  Date \tab 2016-09-17 \cr
  License: \tab GPL (>= 2) \cr
}
}
\references{
\tabular{l}{
         Fan J, Feng Y, Wu Y. \emph{Network exploration via the adaptive LASSO and SCAD penalties.} The annals of applied statistics, 2009, 3(2): 521. \cr
         Friedman J, Hastie T, Tibshirani R. \emph{Sparse inverse covariance estimation with the graphical lasso.} Biostatistics, 2008: 9.3: 432-441. \cr
         Lee W, Liu Y. \emph{Joint estimation of multiple precision matrices with common structures.} Journal of Machine Learning Research, 2015, 16: 1035-1062. \cr
         Li H, Gui J. \emph{Gradient directed regularization for sparse Gaussian concentration graphs, with applications to inference of genetic networks.} Biostatistics, 2006, 7(2): 302-317. \cr
         Lyu X, Sun W, Wang Z, Liu H, Yang J, Cheng G. \emph{Tensor Graphical Model: Non-convex Optimization and Statistical Inference.} IEEE transactions on pattern analysis and machine intelligence, 2019, 42(8): 2024-2037.
    }
}
\author{
\tabular{l}{
  Xiang Lyu, Will Wei Sun, Zhaoran Wang, Han Liu, Jian Yang, Guang Cheng. \cr
  Maintainer: Xiang Lyu <xianglyu@berkeley.edu>
}
}
