% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varcor.R
\name{varcor}
\alias{varcor}
\title{Variance Correction of Sample Covariance of Residuals}
\usage{
varcor(data, Omega.list, k = 1)
}
\arguments{
\item{data}{tensor object stored in a m1 * m2 * ... * mK * n array, where n is sample size
and mk is dimension of the kth tensor mode.}

\item{Omega.list}{list of precision matrices of tensor, i.e., \code{Omega.list[[k]]} is precision matrix 
for the kth tensor mode, \eqn{k \in \{1 , \ldots, K\}}{ 1 <= k <= K }. 
Elements in \code{Omega.list} are true precision matrices or estimation of the 
true ones, the latter can be output of \code{Tlasso.fit}.}

\item{k}{index of interested mode, default is 1.}
}
\value{
A scalar of variance correction for the kth mode.
}
\description{
Generate variance correction term of sample covariance of residuals described in \href{https://arxiv.org/abs/1609.04522}{Lyu et al. (2019)}.
}
\details{
This function computes variance correction term of sample covariance of residuals and
 is utilized to normalize test statistic into standord normal, see \href{https://arxiv.org/abs/1609.04522}{Lyu et al. (2019)}.
}
\examples{

m.vec = c(5,5,5)  # dimensionality of a tensor 
n = 5   # sample size 
k=1 # index of interested mode
lambda.thm = 20*c( sqrt(log(m.vec[1])/(n*prod(m.vec))), 
                   sqrt(log(m.vec[2])/(n*prod(m.vec))), 
                   sqrt(log(m.vec[3])/(n*prod(m.vec))))
DATA=Trnorm(n,m.vec,type='Chain') 
# obersavations from tensor normal distribution
out.tlasso = Tlasso.fit(DATA,T=1,lambda.vec = lambda.thm)   
# output is a list of estimation of precision matrices

rho=covres(DATA, out.tlasso, k = k) 
# sample covariance of residuals, including diagnoal 
varpi2=varcor(DATA, out.tlasso, k = k)
# variance correction term for kth mode's sample covariance of residuals

}
\seealso{
\code{\link{varcor}}, \code{\link{biascor}}, \code{\link{covres}}
}
\author{
Xiang Lyu, Will Wei Sun, Zhaoran Wang, Han Liu, Jian Yang, Guang Cheng.
}
